+++ src/middleware.ts
@@ -51,5 +51,5 @@ export function middleware(request: NextRequest) {
 }
 
 export const config = {
-  matcher: ["/((?!_next/static|_next/image|favicon.ico|assets/|icons/).*)"],
+  matcher: ["/((?!_next/static|_next/image|favicon.ico|favicon.png|assets/|icons/).*)"],
 };